added GAME solver when Lindblad operators are time independent
The GAME equation is Dragomir Davidović, Quantum 4, 326 (2020) and in
Donghyun Jin, Grihith Manchanda, and Dragomir Davidovi ́c,
arXiv:2105.09766.

GAME is in Lindblad form, but the Lindblad operators and Lamb shift
take a particular form in terms of a so-called "timed-spectral density."
In the 'time-independent case,' i.e., when the system Hamiltonian H_S
and the bath/interactions are constant, the resulting Lindblad operators
are also constant. We have programmed this case here.

Note: To generalize to the time-dependent case naturally, we choose
to write a lower-level solver for GAME than is used in the other parts
of the HOQST (or OpenQuantumTools.jl) package. That is, we actually
perform numerical integration/ differential "by hand" instead of using
outside packages. This leads to a lack of consistency in the implementation
obviously, but also slightly in terms of the exposed interface. It's
a future issue to figure out how to resolve this.